+++ tailwind.config.js
@@ -1,0 +1,8 @@
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: ['./inertia/**/*.{js,ts,jsx,tsx}', './resources/views/**/*.edge'],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+}