+++ CHANGELOG.md
@@ -1,9 +1,23 @@
 # Change Log
 
+## [UNRELEASED]
+
+## [0.3.0] - 2024-05-07
+
+- Add `TailwindCSS` to the project.
+- Setup `Redis` to handle session storage.
+- Add `docker-compose.yaml` to the project to handle PostgreSQL and Redis in development environment.
+- Setup `OAuth2` authentication's controller and routes (Twitch for now).
+- Add i18n handling on frontend with `react-i18next`.
+- Setup default layout to pages.
+
 ## [0.2.0] - 2024-05-01
+
 ### Added
+
 - ~~Add i18n handling to the project with `next-intl`.~~
 - Switch project from `NextJS` to `AdonisJS`.
 
 ## [0.1.1] - 2024-04-25
+
 ### Added